Update CSV file columns
Motivation:

Modify the columns of the CSV in preparation for the transition to the new structure of PaNET.

Modification:

Added 'Equivalent', 'Super Property', 'Domain', 'Range', 'Characteristic' columns.

Removed 'Class Type' column as it is an optional column.

Replace the class template strings under 'Parent IRI' columns, which was 'CI', with 'SC %'. This is the preferred class template string according to the ROBOT documentation:
http://robot.obolibrary.org/template#template-strings

Result:

The class template strings have been updated, and the necessary additional columns have been added, all in preparation for the new structure of PaNET.